refactor: extract DDL, co-activation, and indexes from knowledge_store.py into storage/
knowledge_store.py: 4,603 → 3,671 lines (-20%)
New: storage/ddl.py (670), storage/co_activation.py (349), storage/indexes.py (175)
